fb: sirf: fix the issue of saving and restoring framebuffer
any memblock removed will be not managed by linux, then there
is no page, and pfn becomes invalid.
this patch makes the framebuffer becom reserved instead of being
removed, then linux will manage it and save. so we can't ioremap
it after we move to use memblock_reserve. we need to move to
phys_to_virt as the framebuffer memory is not IOMEM any more.
Jiansong/renwei:
will you only make the 1st framebuffer which want to be saved
reserved, and make other removed?

Signed-off-by: Barry Song <Baohua.Song@csr.com>