sd0 pm: sirf: add broken-cd and non-removable
Since sd0 refer to boot media and no card detect and
removable support anymore. So here involve broken-cd
and non-removable to deal with.

platfrom_resume used to set sd controller with proper
timeout and delay value which is lose after suspend.
This is known bug and additional to take care of.

Now sleep wakeup with sd0 boot is normal.

Change-Id: Ia7219973318a4d5fbd3cd3fe958095e0b8593800
Signed-off-by: Bin Shi <Bin.Shi@csr.com>