arm:defconfig:sirf:enable CMA to get memory from contiguous area
this makes framebuffer and other drivers to get memory from
bitmap-based CMA instead of system buddy. and then the big-
resolution LCD will always get its framebuffer.

as we will not alloc/free the area dynamically, it supposes
that performance will not decrease due to write-back dirty
pages to disk.

Signed-off-by: Barry Song <Baohua.Song@csr.com>