sirf_memc_monitor:misc: create the memory controller monitor driver
The memory controller monitor(memcmonitor) is the special hardware
which monitor bandwidth, latency, timeout, address hit of each port
between the masters and the memory controller. It is constructed by
four child monitors do the jobs.

This patch enables all the child monitors through the proc filesystem.
The root dir is /proc/sirf_memcmon.

Enable the driver when configging the kernel:
	Device Driver -> Misc devices -> SirfSoC memory controller
		monitor driver

[Bandwidth Monitor]
The bwmon is the switch and the bwinfo exports the info cattached by
the bandwidth monitor to userspace.
Usage:
	Start bandwidth monitor:
	$echo 1 > /proc/sirf_memcmon/bwmon
	Stop bandwidth monitor:
	$echo 0 > /proc/sirf_memcmon/bwmon
	Then get the monitor info by:
	$cat /proc/memcmon/bwinfo

[Latency Monitor]
The latmon is the switch and the latinfo exports the info cattached by
the latency monitor to userspace.
Usage:
	Start latency monitor:
	$echo 1 1000 2000 > /proc/sirf_memcmon/latmon
	The param1 is the switch vaule and the other two are
	rlat_threshold and wlat_threshold.
	Stop latency monitor:
	$echo 0 0 0 > /proc/sirf_memcmon/latmon
	Then get the monitor info by:
	$cat /proc/sirf_memcmon/latinfo

[Timeout Monitor]
The tomon is the switch and the toinfo exports the info cattached by
the timeout monitor to userspace.
Usage:
	Start timeout monitor:
	$echo 1 100 > /proc/sirf_memcmon/tomon
	The param1 is the switch value and the other one is timeout
	threshold value.
	Stop timeout monitor:
	$echo 0 0 > /proc/sirf_memcmon/tomon
	Then get the monitor info by:
	$cat /proc/sirf_memcmon/toinfo

[Address Monitor]
Export one proc file /proc/memcmon/addrmon for user to control the
address monitor. Set 4 ranges address(physical) to it and the monitor
will produce a interrupt when the one of address range has been hit.
The detail message should be print out and the address monitor stopped.
Usage:
	Start address monitor:
	$echo 1 0x<saddr0>-0x<eaddr0> 0x<saddr1>-0x<eaddr1> \
	0x<saddr2>-0x<eaddr2> 0x<saddr3>-0x<eaddr3> \
	> /proc/sirf_memcmon/addrmon
	The param1 is the switch followd by 4 address ranges.
	Stop address monitor:
	$echo 0 > /proc/sirf_memcmon/addrmon

Change-Id: I3070f59bcf877d629788975ae13a6d7b2161c4a2
Signed-off-by: Ye He <ye.he@csr.com>