pwm: sirf: correct state backing from hibernation
Backing from hibernation image, there are 2 isuue about pwm state and
its clock enable count.

clock issue,
clk_prepare_enable() is no need to invoke because clock of pwm
controller is already enabled, that is, clock enable count is greater
than 0 during making hibernation image. It will lead to wrong clock
enable count if the clock is enabled again, so implement restore member
of dev_pm_ops that do not invoke clk_prepare_enable().

pwm state,
Also in the same way, some channel is enabled but not enabled in fact,
so it should be enabled for use of consumer.

Change-Id: I2da6b6e615fd56387bddfec780fba6a1b261e0ce
Signed-off-by: Huayi Li <huayi.li@csr.com>