serial: sirf: correct check condition for fetch dma buffer into tty
In rx dma-callback it call tasklet_schedule, if the tasklet
be scheduled after all the dma-callback in the rx dma channel,
current check condition in the tasklet will not do fetch dma
buffer into tty because tx_issued equal with tx_completed,
so as timeout tasklet does.

so check if fetch the whole dma buffer into tty according
the status of transactions in rx dma channel.

Change-Id: Iead9e00b74fb90ce05e9dfd80fb6bb14ec1a9a60
Signed-off-by: Qipan Li <Qipan.Li@csr.com>