ARM: GIC: provide the ability to config the interrupt group of SPI32-63
All SGI will be non-secure as we don't use SMP in backend Linux. and we
define SPI64~max as non-secure as well. so here provide a configuration
to make we have the ability to define the IGROUP of SPI32-63.
Default, we enable timer and uart1 of vexpress as secure interrupt.

Signed-off-by: Barry Song <Baohua.Song@csr.com>