ARM: sirf: add lookup table for some device
Some application has specific requirements on the naming of devices,
e.g. the name of backlight tuning file will be changed as device
tree changed, which may be confusing as below,

/sys/class/backlight/sirf-backlight.19/brightness or
/sys/class/backlight/sirf-backlight.20/brightness or
/sys/class/backlight/sirf-backlight.xx/brightness.

So provide a device name lookup table to of_platform_populate() to
bind some devices with specific name, e.g. the name of backlight
tuning file will have a specific name as below,

/sys/class/backlight/sirf-backlight/brightness.

Other devices can be added to this lookup table if have the similar
requiremnet

Change-Id: If2fab0772dce32affb6acf8ffb3b617ddb9d32e6
Signed-off-by: Huayi Li <huayi.li@csr.com>