ARM: sirf: reset_controller: move to subsys_init
Arnd:
Generally the hack that people use when things fail to work correctly
is to hijack an earlier initcall level, e.g. subsys_initcall rather
than using module_init. Please first see if that can solve your problem.
If it doesn't help, please resubmit your current patch with a detailed
description in the code on why you have to do it like this.

Signed-off-by: Barry Song <Baohua.Song@csr.com>