ARM: vexpress: load backend kernel and dtb together
here we wrap backend kernel and dtb as a package and let csrvisor
jump to the base address of this package directly.
so we enable ARM_APPENDED_DTB. the ca9 board with dtb will hang
during boot, so we still use the legacy mach init and don't want
to pay attention to vexpress platform issues.

Signed-off-by: Barry Song <Baohua.Song@csr.com>