pwm: sirf: do not use step up mode
In PWM Output Enable Register(PWM_OE),
bit[7...13] denotes transition mode of channel 0 to 6 of pwm.
0: step up mode
1: normal mode

For backlight pwm consumer, step up mode is used to dim backlight
when device suspends, while most of applications can do this with
normal mode, and for other pwm consumers, this feature is useless,
so it is no need to set this mode.

Change-Id: I6266ec2cd2e8928947caff79e1957d46ac5af68a
Signed-off-by: Huayi Li <huayi.li@csr.com>