driver:remoteproc:supports private data of virtio remoteproc device
Sometimes, virtio remoteproc device may need private data. This private
data can be device configuration or child device descriptor, and could not
be stored in the MMIO space.

For example: virtio rpmsg bus device. RPMSG is a bus device, it may contain
some child device descriptors. If the number of child devices is large, the
table of descriptor can be huge. It's too huge to be stored in MMIO space.
So we defined 2 MMIO offset in remoteproc:

MMIO_PRIV_DATA:The memory address of virtio rproc device's private data.
For RPMSG example, this offset stores the address of the table of child
devices.

MMIO_PRIV_SIZE:The size of the private data. The unit can be byte or array
size, etc. This can be defined by device and driver.
For RPMSG example, this offset stores the array size of descriptors.

Change-Id: Ia4b246ff5c4326dc9ac848ebed1341bb14745b15
Signed-off-by: Wei Chen <Wei.Chen@csr.com>