arm:vexpress:add csrvisor remote processor device into platform device list
csrvisor remote processor device (csrvisor-remoteproc) is a IPC device
which takes the responsibility of communicating with backend and frontend OS.

csrvisor-remoteproc can be found in both sides, we have emulated it as a
real IPC device, it has io memory range, interrupts and act as a device bus.

we should register csrvisor-remoteproc into platform device list make
the driver has better compatibility with ARM architecture.

Change-Id: I750003f2d8d6e7ef3efda8b61bb3e04067d74240
Signed-off-by: Wei Chen <Wei.Chen@csr.com>