driver:remoteproc:allocate default vring space for device doesn't support dynamic vq
If the device support dynamic virtqueue, the rproc will allocate a
space that will hold max to 32 virtqueue for device. But, if the
device doesn't support dynamic virtqueue, the rproc forget to allocate
default virtqueue size 2 for such device. This will cause memory overflow
if there are multiple rproc_dev on rproc bus.

Change-Id: I1b5e45b6fbf5a63f042b8c30d9d25944f9ec7a46
Signed-off-by: Wei Chen <Wei.Chen@csr.com>