sirf: fs :bugfix: kernel panic when run r/w on removeable block and do suspend
(sirf APBU-8340) When suspend, it will plug out removeable block device
and unregister the bdi (bdi->dev = NULL)at the end of syncing the file system.
But when some threads writing on these devices and it is not freezing,
if dirty pages exceed the limit, it will call bdi_start_background_writeback
modify the timer delay time to 0, The bdi worker function is called.
But the dev is NULL,  read the dev name cause panic.

Check the bdi->dev if it is null every time to fix it.

Even the these threads make the bdi worker fuction run after the bdi
is unregister. The bdi_unregister clear the bdi list,
The worker function do nothing.

[ 1917.544181] Unable to handle kernel NULL pointer dereference at virtual address 0000002c
[ 1917.551830] pgd = c0004000

[ 1917.554465] [0000002c] *pgd=00000000
[ 1917.557868] Internal error: Oops: 5 [#1] PREEMPT SMP ARM
[ 1917.563154] Modules linked in: vxdkernel(O) sirfsocfb(O) pvrsrvkm(O)
[ 1917.569493] CPU: 0 PID: 2678 Comm: kworker/u2:2 Tainted: G           O 3.10.16 #51
[ 1917.582068] task: c4d31400 ti: d4410000 task.ti: d4410000
[ 1917.587447] PC is at bdi_writeback_workfn+0x1c/0x154
[ 1917.592397] LR is at process_one_work+0x218/0x334
[ 1917.597078] pc : [<c00e76b4>]    lr : [<c00341a0>]    psr: 60000013
[ 1917.597078] sp : d4411ed0  ip : 00000000  fp : 00000000
[ 1917.608524] r10: d4410028  r9 : 00000000  r8 : d8928bc0
[ 1917.613730] r7 : d8928c38  r6 : d902bc00  r5 : d8928a90  r4 : d8928bcc
[ 1917.620236] r3 : 00000000  r2 : d4411ee0  r1 : 00000002  r0 : c051d221
[ 1917.626743] Flags: nZCv  IRQs on  FIQs on  Mode SVC_32  ISA ARM  Segment kernel
[ 1917.634029] Control: 10c53c7d  Table: 0b8c804a  DAC: 00000015
[ 1917.639752]

Change-Id: I39762f61675bfa515596d61feaa4ae699e15ecfe
Signed-off-by: Minda Chen <Minda.Chen@csr.com>