vip: sirf: replace reserved memory with cma in vip
vip's continous memory is reserved in init which limit vip to be built-in.
So if vip is built as module, cma has to be used.
Solution: add a fake cma device and declare cma area in init, then wrap
the cma area to vip device. It's transparent for vip to allocate continous
buffer by using dma_alloc_coherent.

Change-Id: Id37b8b879c1393811edeac552dbc074b07c85544
Signed-off-by: Lihui Wen <Lihui.Wen@csr.com>