ARM: vexpress: map the frontend OS to a backend thread
we involve a L4 or ADEOS/XENOMAI similar ways to map the frontend OS
to a thread in backend.
Then even the backend CPU is 100% used by tasks, frontend OS still
have timeslots to run.
we think high CPU loading in backend OS is a bug, but it still makes
sense for us to provide a way to let frontend run in this case.

Change-Id: I6044e0d134877a3589d7502e2807e6cb9d85a479
Signed-off-by: Barry Song <Baohua.Song@csr.com>