drivers:remoteproc:Fix the memory overflow of sirf_remoteproc in single OS
The private driver data for Secure OS rproc will be disable in single
OS environment, but the index in of_device_id table doesn't consider
this change. So, in single OS the rproc driver will get mistake
private data which is outside the table.

In dual OS environment, we fix the fifo channel size to 0x10000. This
require the share memory size is greater than 0x10000 * 2. But in rproc
for M3 and KAS, the share memory size should be less than or equal to 64KB.

Now we place the fifo channel size to private data. Each different rproc
could configurate its fifo channel size to fit its requirement.

Change-Id: Ie02253a3cd4b20720e66f22e15c0e7622b9dc511
Signed-off-by: Wei Chen <Wei.Chen@csr.com>