drivers:rpmsg:remove the dependence of CSRVISOR_REMOTEPROC
In previous design, rpmsg has been splited into backend and
frontend two parts. And they are depended on CSRVISOR_REMOTEPROC.

In single OS environment, rpmsg is always acts as frontend driver,
and it will not depend on CSRVISOR. If remoteproc is on and secure
mode is disable, the rpmsg frontend can be select by other module.

Change-Id: Ifc332df2a69b4f95ad5c462c461b828f6e6e3a2c
Signed-off-by: Wei Chen <Wei.Chen@csr.com>