drivers:remoteproc:Using macro to replace hardcode registers' offset
Currently, the IPC interrupts' trigger and clear registers
are hard code in the ipc hwinfo table. If the registers'
offset are modified by hardware design, we have to modify
hwinfo table frequently, and it's very unreadable.

So we use the registers' macro definition to replace hardcode.
It's better for read and modification.

Change-Id: I3b7c71e352b1248600f71c01ee8c28009c1de0b2
Signed-off-by: Wei Chen <Wei.Chen@csr.com>