sdhci : sirf: add sirf tuning fuction (cmd 19)
Add manual tuning function in sirf atlas7 soc.
It is mainly used for the UHS-I SD card working SDR50
SDR104 mode.

The tuning principle can be seen in SD spec part1 v3.01 4.2.4.5
(tuning command).

SD host send the cmd19 and set the delay value(0-127).
and the sdcard return 64 bytes data. If the data is same with
the tuning data. The delay value is valid. Execute this commmand
128 times. And calculate the longest window of the valid values.
The value in the middle of this window is the best value.

Change-Id: I4a0734c0fc04a53b43635c441472219112481c00
Signed-off-by: Minda Chen <Minda.Chen@csr.com>