drivers:pinctrl:enable A/D mode check before atlas7 pin set mux
In atlas7, some pins support analogue and digital input.
Before pin set mux, we have to check whether this pin
supports A/D mode.

If this pin doesn't support, we'll skip the A/D mode convert,
and set the target mux func for this pin.

If this pin supports. If the target mux func is analogue, we
just set this pin to analogue mode. If the target mux func is
not analogue, we have to set this pin to digital mode and then
set the pin's mux func to target mux func.

Change-Id: I22491ef7449f144c4843649f4979a5206937accd
Signed-off-by: Wei Chen <Wei.Chen@csr.com>