arm:dts:reserved a fixed memory region for atlas7 DMA area
The DMA memory should be shared between M3 and A7 cores,
for example: rpmsg for M3-A7. But the memory address can
be accessed by M3 is limited to 384MB.

On Atlas7 Linux the default DMA area is reserved in the
last 16MB memory of DDR RAM. If the DDR RAM size is larger
than 384MB, the M3 could not access the DMA memory allocated
by Atlas7 Linux.

Now we reserved the DDR memory from 0x4EC00000 to 0x4FC00000
for DMA contigous. Both M3 and A7 can access this region.

Change-Id: I60c96c71b366d37d3de18d689280656ef6c9a03f
Signed-off-by: Wei Chen <Wei.Chen@csr.com>