drivers:remoteproc:adjust the order of remoteprocs of Atlas7
In the previous setting, if we enable the DualOS
the NS-M3 rproc index will be changed from 0 to 8.

This change will affect the index of hwspinlocks
that are used by rproc software fifo. And we have
to modify the M3 code to sync this change.

Now, we adjust the index order of remoteprocs of
Atlas7 from zero, if we enable the DualOS, it doesn't
affect the hwspinlocks, and we haven't to modify
M3 code.

Change-Id: I60a7833986841bd9f24f57ebab0cefd92fd25849
Signed-off-by: Wei Chen <Wei.Chen@csr.com>