media:mjpeg:add mjpeg driver for a7
JPEG core is responsible for compressing or decompressing an image. The HW handles the ECS
of the bit-stream. Both compression and decompression can be done in YUV422/YUV420 formats.
The core can be used to create MJPEG videos by compressing each frame as JPEG and combining
all frames in a video container.

Change-Id: If5a9d9ca4a50c4cb4071ce88fa36aaa231f1e22a
Signed-off-by: Lily.Li <Lily.Li@csr.com>