ASoC: atlas7-iacc: Use the hrtimer to simulate period interrupt
The atlas7 loop dma mode only support the period length is half
of the buffer length, So if the audio pcm driver use period length
is not half of buffer length, the dma driver, the dma driver
non-acceptance. So in the audio pcm driver, use a hrtimer to
simulate the small period piece interrupt.

Change-Id: If9760bb10441090a597ed5fae81fac6a530f5032
Signed-off-by: Rongjun Ying <rongjun.ying@csr.com>