vdss: sirf: replace the get_default_panel with get primary/secondary device.
As only enabled panel is registered to vdss core, the default panel doesn't
need anymore. This patch removes the default panel from vdss. Besides this,
the framebuffer will fetch primary or secodary panel from the registered
panel list according to the display alias list in the dts file. The logic
is applicable to the multi display case.

Change-Id: I652c30bbbf6401017d37c65d76d460963e19a91f
Signed-off-by: Renwei Wu <renwei.wu@csr.com>
Signed-off-by: Jiansong Chen <jiansong.chen@csr.com>