virtio: remove dual-os support and csrvisor remoteproc
this causes the sirf-remoteproc depends on core level changes even for
the remoteproc with M3.
and it is not mergable for 3.19 and coming kernel. we are losing maintain
for it.
if oneday we need it again, apply it to 3.18. if oneday, we need it on newer
kernel, rebase and test are needed.

Change-Id: Ic32a385e3f307a8d9de5a2176692c065b10edd98
Signed-off-by: Barry Song <Baohua.Song@csr.com>