dts: CSRatlas7: add fixed regulator for usb0
The behavior of the usb0.drvvbus doesn't confirm to the OTG spec. It will
automatically fall down when switching from Host to Device. It is solely
controlled by the working mode of the usb controller.
We correct this by configuring the function of the usb0.drvvbus to gpio and we
simulate this gpio as a fixed regulator to control the on/off of the usb0 vbus
power.

Change-Id: If23bfebaa81adfa25d9f6968a02a308ad5c65a1e
Signed-off-by: Rong Wang <Rong.Wang@csr.com>