scripts: dtc: define CONFIG_ANDROID if we enable Android
this is a workaround for making the dts maintain easier. we have two
different BT/WiFi modules on EVB board, for Android, we want to use
Amber, for Linux, we want to use internal A7CA.

once we have refined the dts arch and support the sync between multi
ple dts files, revert this.

Change-Id: Ie1f1260255699b77bf8694475e9aca2441afdad8
Signed-off-by: Barry Song <Baohua.Song@csr.com>