nanddisk: do not use buffer of user space in ioctl
The struct nanddisk_ioctl include 2 buffers(in_buf and out_buf), in which
the nanddisk command parameters(io, config, power on/off, etc.)are saved,
but they cannot be accessed directly because they are in user space.

Also, there is a data buffer in the nanddisk command parameter(in_buf) which
cannot be accessed directly when doing read or write ioctl. The relation
of them is looked as below,

nanddisk_ioctl -> in_buf(io parameter) -> data_buf

Change-Id: Ibd89fd2fd02bba8b1764b740dcc249332ecfc50c
Signed-off-by: Huayi Li <huayi.li@csr.com>