drivers:remoteproc:support predefined vq notifyid
In current rproc use scenario, the vq notifyid
will be allocated by Linux, and write back this
notifyid to RSC table to tell DSP.

But in some scenarios, we might meet the DSP
predefined the vq notifyid. In these scenarios,
the rproc have to read predefined vq notifyid
from RSC table, and then request a same notifyid
in Linux side.

Change-Id: I8934ca3efea2abfd36a74430a5f0218942663955
Signed-off-by: Wei Chen <Wei.Chen@csr.com>