drivers:remoteproc:support always on remote processor
In current rproc use scenarios, the Linux side controls
the lifecycle of the remote processors, just like boot,
start, shutdown.

But in some scenarios, we require the remote processors
to be powered and booted before Linux bring up. In
additional, we might have require the remote processors
could not be shutdown, if it would be shutdown, the Linux
should be shutdown too.

In above scenarios, we have make remoteproc to support
always on remote processor. This kind of remote processor
control its lifecycle by itself.

Change-Id: Id3683de9894b8797c344edc078cbb720f184f659
Signed-off-by: Wei Chen <Wei.Chen@csr.com>