drivers:remoteproc:support atlas7 M3 remote processor
CSRatlas7 has a M3 core as a remote processor.
We use the remoteproc framework to communicate
with this Mcore.

A7 and M3 core reserved a range of memory to share
data. This memory area is cached.
We have to make the share memory un-cached, but we
can't reserve a 64Kbytes un-mapped memory region in
DTS (The mini size of un-mapped region is 1MB).
So we have to call: dma_common_contiguous_remap
to remap share memory to un-cached DMA memory.

This driver depends on Atlas7-IPC, Atlas7-hwspinlock,
so we enable these two options in prima2_defconfig.

Change-Id: I11469fc8840b89632afb9fbc3e1a4bfc6a6da93e
Signed-off-by: Wei Chen <Wei.Chen@csr.com>