mmc:sirf:host Validate SDR50 at 100MHz
We need to add SDHCI_QUIRK2_PRESET_VALUE_BROKEN to disable preset value
usuage since we didn't validate them in our host controller, or else
there may be abnormal in clock switch.
We also need to make small modifcation to sdhci_sirf_get_max_clk function
to round the max clock to multiples of 1Mhz. Or else we may not get 1 divisor
when setting I/F clock. For example max clock is actually 100000096Hz, and target
clock is 100000000Hz, we can only get the divisor of 2 and the actual clock is
50000048Hz.

Change-Id: I28f07c0bf601e2d8109f7ca75005ed351db49a75
Signed-off-by: weijun yang <york.yang@csr.com>