mmc:sdhci:sirf:Reset clock delay after UHS card removal
In case the UHS card is inserted and run in SDR50 mode, clock delay
tunning will be called and a fine tuned value will be set in
SD_CLK_DELAY_SETTING(0x004C)
While this value may be not fitful for HS card. So after the insertion/
removal of a UHS card, the later inserted HS card may not be properly
initialized.
Now we validate the function of SD_CARD_EVENT, in case the card is
removed, the clock delay will be reset.

Change-Id: Ib46b1964d57ebb14a92249ea901c8c6b8764d693
Signed-off-by: weijun yang <york.yang@csr.com>