pm: sirf: fix platform restart delay
currently use  watchdog as workaround for platform
restart, but the watchdog configure sequence might
wrong. which cause restart delay more than 15 sec.
the correct sequence should be:
1.clear the counter
2.set enough match  value
3.enable the counter control
4.enable the watchdog reset feature of the counter
the  reason for  enough match is that, after 3, the counter
start kicking, and before 4 is done, the counter is run at
150M clk, so the match should be large enough.

with the patch the restart delay is less than 1/8.

Change-Id: I5850c35ebf01d1489aaf6ce5327157e9c3ad9a7c
Signed-off-by: Guo Zeng <guo.zeng@csr.com>