vdss: sirf: Entry Evaluation Board supported
In Entry, the 16bit LCD panel is mixing the data lines. This is
why the 16bit panel for dual chip is not trivial. So for entry chip,
we need to take into account the mixing in the chip package and the
mixing in the LCD panel board and set the pad-mux accordingly.

Change-Id: I6be71d55a99c41de2824ad964ee6bd78a77c31a8
Signed-off-by: Dongli Li <Kasin.Li@csr.com>