media: vip: cvd: fix wrong field order issue
When interlaced video input, either rearview case or DVD play case
sometimes the picture quality on the screen is very bad
because of the wrong field order data processed by VPP.

Because chip IP design changes, we can't use VIP_ODD/EVEN_CAPTURE function
before step_B, currently we have to get field ID in the CVD, and pass it
to the VIP, VIP will set the field order information in the v4l2_buf struct
that userspace and VPP will be easy to get.

But there is also a bug in the CVD, though all the signal is stable and locked,
sometime the readout FIDs of the beginning several field are wrong.
We have to skip the beginning 4 fields, then read the FID is.

To be convenient for "replaced by VIP_ODD/EVEN_CAPTURE function in future" and
"VIP fifo overflow handle", CVD outputs top->bottom fixed field order frames.

Change-Id: I439a112a3bfbb46762549e6196dc02689a9fe0d4
Signed-off-by: as64 <Andy.Sun@csr.com>