media: cvd: Fix vip suspend cause hang
The following error occurred when suspend cvd. It is caused by trying to
access hardware while clock is off.

[   16.490000] sirfsoc-cvd 10db2000.cvd: cvd_pm_suspend
[   16.490000] sirfsoc-vip 10db0000.vip0: vip_pm_suspend
[   16.500000] NoC: err:        power disconnect
[   16.500000] NoC: ID: A7
[   16.500000] NoC: Opc:        read
[   16.510000] NoC: Addr        10db0414
[   16.510000] NoC: Len 00000003

In fact CVD and VIP are in the same module in hardware side, They share
one clock source. I added the same clock source in both CVD and VIP to
ensure that the clock is closed by the final clk_disable_unprepare.

Change-Id: I50fbaffb1bffcb70b28cfcbcdc58458c0c8acec7
Signed-off-by: Dongli Li <Kasin.Li@csr.com>