media: mmc: tunex: add an interface to set the timer interval
add an interface to set the hrtimer interval for sdr daemon, it can set
low interval to check the dma buffer when dab or fm mode(high data rate)
and set high interval for am mode(low data rate).

It is just for test different setting for different data rate to help
tuning the performance now.

Change-Id: I4b70e3fe4cd30dadad8220e3f89e493a9b771620
Signed-off-by: Yonghui Zhang <yonghui.zhang@csr.com>