usb: chipidea: reuse the platform_data to store the ci info
Normally we save the driver data to the driver_data of the device struct.
But for the dual-role chipidea device, the driver_data will be used to save
the info about ehci, so we can't use it any more.
While the platform_data would not be accessed after the device is added to
system after the probe process, so it's safe to reuse it here.

Change-Id: Id218e84e413211bad8e8a2f0b078e51790556c79
Signed-off-by: Rong Wang <Rong.Wang@csr.com>