iio: adc: sirfsoc_adc: use mode bits to sync adc operations with M3
Atlas7 and M3 may share the ADC concurrently.
use mode bits in control1 register to ensure if
the adc is busy, If the mode bits is 0, the adc
could be use, otherwise, the adc is using by the
other cpu(A7 or M3), and need discard this
operation and request again.

Change-Id: I05588f21a9125cb41cf3a15b8ccb18695a9dd520
Signed-off-by: Guoying Zhang <Guoying.Zhang@csr.com>