media: rearview: use itself workqueue not shared queue
After triggering rearview, all the works will be done in the worker which
locates in a shared workqueue, if user disconnect camera, the worker in the
shared workqueue will block(wait for camera connection) and won't return,
in that case othet workers in that shared workqueue will be influenced.

So we create a private workqueue for rearview using.

Change-Id: I0edefd5dc4a29de1a42df7e9b58512b7236a4af6
Signed-off-by: as64 <Andy.Sun@csr.com>