i2c: i2c-sirf: modifiy divider range to make i2c bus clock more accurate
Since i2c is open drain interface that allows the slave to
stall the transaction by holding the SCL line at '0', the RTL
implementation is waiting for SCL feedback from the pin after
setting it to High-Z ('1'). This wait adds to the high-time
interval counter few cycles of the input synchronization
(depending on the SCL_FILTER_REG field), and also the time it
takes for the board pull-up resistor to rise the SCL line.
For slow SCL settings these additions are negligible,
but they start to affect the speed when clock is set to faster
frequencies.
So use the actual test, to get the suitable param in different
ranges of the i2c bus clock frequency, to make the SCL frequency
more accurate.

Change-Id: Ie77b2a8ba5b3e30b2e032c026ca58aae4d9116d0
Signed-off-by: Guoying Zhang <guoyingz@codeaurora.org>