clk: sirf: update reset controller for step b
The reset controller have been changed in atlas7 stepB soc. We have to
update the driver to support step B chip.

Changes:
    The bits of these registers were redistributed into new parameters
    based on functionality.
    The new registers are designated CLKC_RSTC_<func>_SW_RST_SET/CLR/STATUS

    CLKC_RSTC_*_SW_RST_SET
    CLKC_RSTC_*_SW_RST_CLR
    CLKC_RSTC_*_SW_RST_STATUS

TODO: some of bit in reset register might be missing in hw design, for example,
dmac3 bit 6 in reset does not exist as clk leaf, so as our current design,
the reset index is same as clk index, so some of the reset index is not valid.
will create separate reset table, will put it in the TODO list.

Change-Id: Ia1f3d8f91d0676dba67c0be3e0452f62a0eefc5e
Signed-off-by: Guo Zeng <gzeng@codeaurora.org>